Fixing lots of things

+++ commands/clear.js
@@ -5,36 +5,52 @@ class Clear extends Command {
     }
 
     static action(message) {
+        message.channel.startTyping()
         if (message.member.permissions.has("MANAGE_MESSAGES", true)) {
             let args = message.content.split(' ');
             if (args[1] === "all") {
                 message.channel.fetchMessages()
-                    .then(function (messages) {
+                    .then((messages) => {
                         var siz = messages.size;
                         for (var i = 0; i < siz; i++) {
                             messages.array()[i].delete()
                         }
-                        message.author.send("Des messages ont été supprimé avec succès")
+                        message.author.send(message.i18n.__('clear.all'))
                     })
                     .catch(console.error);
             } else {
-                if (parseInt(args[1]) > 100) {
-                    args[1] = 100;
+
+                if (parseInt(args[1]) < 1) {
+                    var toDel = 0;
                 }
-                message.channel.fetchMessages({limit: parseInt(args[1])})
-                    .then(function (messages) {
-                        var siz = messages.size;
-                        for (var i = 0; i < siz; i++) {
-                            messages.array()[i].delete()
-                        }
-                        message.author.send("**" + siz + "** message(s) viennent d'être supprimé")
-                    })
-                    .catch(console.error);
+                else if (parseInt(args[1]) == "NaN") {
+                    var toDel = 0;
+                }else{
+                    var toDel = args[1];
+                }
+                console.log(toDel);
+
+                if (toDel > 0) {
+
+                  message.channel.fetchMessages({limit: toDel})
+                      .then((messages) => {
+                          var size = messages.size;
+                          for (var i = 0; i < size; i++) {
+                              messages.array()[i].delete()
+                          }
+                          message.author.send(message.i18n.__('clear.some', size))
+                          // message.author.send("**" + siz + "** message(s) viennent d'être supprimé")
+                      })
+                      .catch(console.error);
+                }else{
+
+                  message.author.send(message.i18n.__('clear.some', 0))
+                }
             }
         } else {
-            message.author.send("Sorry man man man ");
-        }
-
+            message.author.send(message.i18n.__('clear').permissions_invalid)
+        }        
+        message.channel.stopTyping();
     }
 }
 module.exports = Clear
